requirements.txt
accelerate @ file:///workspace/MiniGPT4-on-MLU/wheel/accelerate-0.20.3-py3-none-any.whl
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.5
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
braceexpand==0.1.7
cached-property==1.5.2
cambricon-dali @ file:///torch/cambricon_dali-0.8.0-py3-none-manylinux2014_x86_64.whl
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cpplint==1.6.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.6
DLLogger==0.1.0
entrypoints==0.4
exceptiongroup==1.1.2
fastapi==0.101.1
fastjsonschema==2.17.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gradio==3.34.0
gradio_client==0.2.6
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
isort==5.11.5
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.4
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
librosa==0.8.1
linkify-it-py==2.0.2
llvmlite==0.39.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.56.4
numpy==1.21.6
omegaconf==2.3.0
opencv-python==4.7.0.72
orjson==3.9.5
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pluggy==1.2.0
pooch==1.7.0
portalocker==2.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pydub==0.25.1
Pygments==2.15.1
pylint==2.17.4
pyparsing==3.1.1
pyrsistent==0.19.3
pytest==7.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
regex==2023.8.8
requests==2.31.0
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.2.0
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
starlette==0.27.0
svgwrite==1.4.3
terminado==0.17.1
threadpoolctl==3.1.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch @ file:///torch/wheel_py37/torch-1.9.0-cp37-cp37m-linux_x86_64.whl
torch-mlu @ file:///torch/wheel_py37/torch_mlu-1.15.0_torch1.9-cp37-cp37m-linux_x86_64.whl
torchvision @ file:///torch/wheel_py37/torchvision-0.10.0a0%2B300a8a4-cp37-cp37m-linux_x86_64.whl
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.0
Tree==0.2.4
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.7.1
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
wrapt==1.15.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.15.0
